Add files via upload
Because of the issues of having in the main branch after first time doing the cookies, rolled back from the main branch before introducing cookies and redid everything and retested and figured out why the username was not getting passed into the views. Fixed the issue and it is now working as attended and now pushing it this side branch.